Crate pallet_robonomics_lighthouse

Lighthouse is a block author in robonomics parachain.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

InherentDataProvider

The thing that the outer node will use to actually inject the inherent data

Enums

InherentError

Constants

INHERENT_IDENTIFIER

Lighthouse inherent identifier

Type Definitions

InherentType

The type of data that the inherent will contain. Just a byte array. It will be decoded to an actual account id later.